services/stream-processor-service/requirements.txt
fastapi
uvicorn
kafka-python
clickhouse-driver
redis
prometheus-client
